pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "falcon-lang-sayan"
version = "4.8.1"
description = "Secure, AI-powered language ."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-genai",
    "requests"
]

[tool.setuptools]
py-modules = ["falcon_engine", "fpm"]

[project.scripts]
falcon = "falcon_engine:main"

